requirements.txt
numpy==1.22.4
Pillow==9.2.0
scikit_learn==1.1.2
scipy==1.8.1
torch==1.12.1
torchvision==0.13.1
